+++ source/index.js
@@ -46,7 +46,7 @@ function prepareElementRegex (input, addFlags = '') {
  * @private
  * @type {RegExp}
  */
-const elementsRegex = new RegExp('<(?<element>[-a-z]+)(?<attributes>\\s+.+?)?(?:\\/>|>(?<inner>[\\s\\S]+?)<\\/\\k<element>>)')
+const elementsRegex = new RegExp('<(?<element>[-a-z]+)(?<attributes>\\s+.+?)?(?:\\/>|>(?<inner>[\\s\\S]*?)<\\/\\k<element>>)')
 
 /**
  * Get a regular expression for finding a particular element
@@ -57,7 +57,7 @@ const elementsRegex = new RegExp('<(?<element>[-a-z]+)(?<attributes>\\s+.+?)?(?:
  */
 function getElementRegex (element, addFlags) {
 	const { source, flags } = prepareElementRegex(element, addFlags)
-	const regex = new RegExp(`<(?<element>${source})(?<attributes>\\s+.+?)?(?:\\/>|>(?<inner>[\\s\\S]+?)<\\/\\k<element>>)`, flags)
+	const regex = new RegExp(`<(?<element>${source})(?<attributes>\\s+.+?)?(?:\\/>|>(?<inner>[\\s\\S]*?)<\\/\\k<element>>)`, flags)
 	return regex
 }
 
@@ -66,7 +66,7 @@ function getElementRegex (element, addFlags) {
  * @private
  * @type {RegExp}
  */
-const commentElementsRegex = new RegExp('<!-- <(?<element>[-a-z]+)(?<attributes>\\s+.+?)?(?:\\/>|> -->(?<inner>[\\s\\S]+?)<!-- <\\/\\k<element>> -->)')
+const commentElementsRegex = new RegExp('<!-- <(?<element>[-a-z]+)(?<attributes>\\s+.+?)?(?:\\/>|> -->(?<inner>[\\s\\S]*?)<!-- <\\/\\k<element>> -->)')
 
 /**
  * Get a regular expression for finding a particular element
@@ -77,7 +77,7 @@ const commentElementsRegex = new RegExp('<!-- <(?<element>[-a-z]+)(?<attributes>
  */
 function getCommentElementRegex (element, addFlags) {
 	const { source, flags } = prepareElementRegex(element, addFlags)
-	const regex = new RegExp(`<!-- <(?<element>${source})(?<attributes>\\s+.+?)?(?:\\/>|> -->(?<inner>[\\s\\S]+?)<!-- <\\/\\k<element>> -->)`, flags)
+	const regex = new RegExp(`<!-- <(?<element>${source})(?<attributes>\\s+.+?)?(?:\\/>|> -->(?<inner>[\\s\\S]*?)<!-- <\\/\\k<element>> -->)`, flags)
 	return regex
 }
 
@@ -91,7 +91,7 @@ function getCommentElementRegex (element, addFlags) {
  * @returns {string}
  */
 function extractAttribute (attributes, attribute) {
-	const regex = new RegExp(`(${attribute})\\s*=\\s*('[^']+'|\\"[^\\"]+\\"|[^'\\"\\s]\\S*)`, 'ig')
+	const regex = new RegExp(`\\s(${attribute})\\s*=\\s*('[^']+'|\\"[^\\"]+\\"|[^'\\"\\s]\\S*)`, 'ig')
 	let value = null
 	while (true) {
 		const match = regex.exec(attributes)