+++ HISTORY.md
@@ -1,5 +1,9 @@
 # History
 
+## v1.2.0 2018 June 11
+- Fixed element detection when the element had no inner content, e.g. `<el></el>`
+- Fixed attributes being extracted based on their ending, instead of their entirety
+
 ## v1.1.0 2018 June 11
 - Support self-closing elements